third_party/blink/web_tests/http/tests/devtools/console/console-link-to-snippet-expected.txt

Test that link to snippet works.


Running: testConsoleLogAndReturnMessageLocation
name1:1 239
name1:1 42

Running: testSnippetSyntaxError
Line Message was added: snippet:///name2 Error 'Uncaught SyntaxError: Unexpected token '}'':1:1
name2:2 Uncaught SyntaxError: Unexpected token '}'

Running: testConsoleErrorHighlight
Line Message was added: snippet:///name3 Error '42':1:8
Line Message was added: snippet:///name3 Error '-0':2:8
Line Message was added: snippet:///name3 Error 'false':3:8
Line Message was added: snippet:///name3 Error 'null':4:8
name3:2 42
(anonymous) @ name3:2
name3:3 -0
(anonymous) @ name3:3
name3:4 false
(anonymous) @ name3:4
name3:5 null
(anonymous) @ name3:5
